fix: PR Comment 반영
CombinationComment 관련 DTO 분리

Related: #24
